Made player class and the world class World class has been flushed out with to list one of players and the other of places.
3 constructers added to avoid issue later with building the game world

Methods added to return iterators once world is set places and players
should not be removed

Method added to get a place from a player.

ToDo:

method to be added to world so that the player can be gotten from a
place .... may need to be a list.

Deside on the interface of place and methods of player so that methods
can be filled on on World.

Flesh out methods in World.